The application does not offer ability to see full solution table [1/1]
[Problem]
For studying the correct black jack actions, it is useful to view the entire solution table. The application does not
possess a view to do this.

[Solution]
Added a new fragment that shows the solution table. The user can switch between the quiz and the solution table using
the options menu.