Evenly distribute hand types [1/1]
[Problem]
The hand types produced based on an even probability distribution among all the cards will bias hard hands. For the
purpose of memorizing the table, the different hand types should be evenly distributed.

[Solution]
Determine which hand type and then the appropriate cards to populate the field with, creating an even distribution of
hand types.