Clicking solution option twice crashes app [1/1]
[Problem]
Clicking the solution option causes the solution fragment to be added to the FragmentManager again, which throws an
exception, crashing the application. Fragments should only be added to the manager once.

[Solution]
Add the solution fragment during the Activity.onCreate() and have the option selection actions to simply be hiding and
showing the correct fragment.